Implement custom target for Wiktionary, partial redesign
== Custom targets ==
Previously, only title-to-title mapping between namespaces was
allowed (e.g. "Pink Floyd reunites" <-> "Comments:Pink Floyd reunites").

In Wiktionary, besides "Citations" tab, there's also a "Documentation"
tab for templates, thus making this tab's scheme like:
: "Template:Test" <-> "Template:Test/doc"

Not only it has a custom title mapping, but it also targets to the
same namespace!

You can define custom target by adding 'customTarget' to the tab
definition. For example, if you want to have a tab like the Documentation
tab in Wiktionary, you need to define it like this:
: 'customTarget' => '$1/doc',
where $1 will be replaced by the base title.

== Redesign ==
To meet the requirements for the feature described above,
the extension was subject to redesign. Now stock tabs
are re-created. Nothing is changed in the resulting UI.

Change-Id: I09d59425ace71d6fa6cdf409e3f23a31d6d47273